+++ core/compone/__init__.py
@@ -1,3 +1,9 @@
 # ruff: noqa: F401
+
+# This is the public API of compone, nothing else should be used
+# by users other than these APIs.
+# New APIs should be carefully considered before exposing them.
+# compone is intended to be very simple with a small API surface
 from .component import Component
+from .elements import Element, VoidElement
 from .escape import escape, safe